+++ src/main/java/woowacourse/shoppingcart/dto/cartitem/CartItemResponses.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.cartitem;
 
 import java.util.List;
 import java.util.stream.Collectors;
