+++ src/main/java/woowacourse/shoppingcart/dto/order/OrderResponses.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.order;
 
 import java.util.List;
 import java.util.stream.Collectors;
