+++ src/main/java/woowacourse/shoppingcart/dto/product/ProductResponses.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.product;
 
 import java.util.List;
 import java.util.stream.Collectors;
