+++ src/main/java/woowacourse/shoppingcart/dto/product/ProductSaveRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.product;
 
 import woowacourse.shoppingcart.domain.Product;
